Fixed bad branch merge
